renamed app.js to server.js and init.js to app.js  🐿 v2.12.3
