+++ Procfile
@@ -1,1 +1,1 @@
-web: node --max-http-header-size=80000 server/init.js
+web: node --max-http-header-size=80000 server/app.js
