--- server/init.js
@@ -1,2 +1,0 @@
-require('./setup');
-module.exports = require('./app');
