+++ test/server/modules/form/api.spec.js
@@ -6,7 +6,7 @@ process.env.MARKETO_CLIENT_SECRET = 'test';
 import sinon from 'sinon';
 import request from 'supertest';
 import { expect } from 'chai';
-import app, { ready } from '../../../../server/app';
+import app, { ready } from '../../../../server/server';
 import nock from 'nock';
 import SuperTester from 'supertest-parse';
